FIX: NORM FIX:  clear_proccess
