+++ srcs/cmd_list.c
@@ -6,51 +6,11 @@
 /*   By: junhyeop <junhyeop@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/31 20:00:39 by junhyeop          #+#    #+#             */
-/*   Updated: 2024/05/27 19:33:21 by junhyeop         ###   ########.fr       */
+/*   Updated: 2024/05/27 21:16:26 by junhyeop         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-void	set_home(t_head *head)
-{
-	t_list	*node;
-	char	*home;
-
-	node = head->data->env->next;
-	while (node)
-	{
-		if (ft_strncmp(node->key, "HOME", 5) == 0)
-		{
-			home = ft_strdup(node->value);
-			head->data->home = home;
-			break ;
-		}
-		node = node->next;
-	}
-}
-
-t_head	*init_head(char **envp, int argc, char **argv)
-{
-	t_head	*head;
-
-	void_argument(argc, argv);
-	printf("Welcome to minishell!\n");
-	set_signal();
-	g_exit_status = 0;
-	head = (t_head *)malloc(sizeof(t_head));
-	head->data = (t_data *)malloc(sizeof(t_data));
-	if (!head)
-		error_msg(1);
-	head->top = NULL;
-	head->size = 0;
-	head->data->original_stdin = dup(STDOUT_FILENO);
-	head->data->original_stdout = dup(STDOUT_FILENO);
-	head->data->env = (t_list *)malloc(sizeof(t_list));
-	set_env(&(head->data->env), envp);
-	set_home(head);
-	return (head);
-}
 
 t_list	*cmd_list_new(t_head *head, char *command)
 {
@@ -58,7 +18,7 @@ t_list	*cmd_list_new(t_head *head, char *command)
 
 	newnode = (t_list *)malloc(sizeof(t_list));
 	if (!newnode)
-		error_msg(1);
+		error_msg(0);
 	newnode->next = NULL;
 	newnode->prev = NULL;
 	newnode->token = make_token(head, command);
@@ -71,7 +31,7 @@ t_token	*token_new(char *command, int flag, int quote_flag)
 
 	new = (t_token *)malloc(sizeof(t_token));
 	if (!new)
-		error_msg(1);
+		error_msg(0);
 	new->prev = NULL;
 	new->next = NULL;
 	new->cmd = command;
@@ -96,10 +56,10 @@ void	ft_token_add(t_token **lst, t_token *new)
 	if (tmp == NULL)
 	{
 		*lst = new;
-		return ; 
+		return ;
 	}
 	if (tmp == NULL && new == NULL)
-		error_msg(1) ;
+		error_msg(0);
 	else
 	{
 		while (tmp->next)
