+++ srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: junhyeop <junhyeop@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/16 17:10:35 by junhyeop          #+#    #+#             */
-/*   Updated: 2024/05/27 20:27:47 by junhyeop         ###   ########.fr       */
+/*   Updated: 2024/05/27 20:56:26 by junhyeop         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
