+++ srcs/utils.c
@@ -6,7 +6,7 @@
 /*   By: junhyeop <junhyeop@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/23 18:17:19 by junhyeop          #+#    #+#             */
-/*   Updated: 2024/05/27 20:28:35 by junhyeop         ###   ########.fr       */
+/*   Updated: 2024/05/27 20:56:42 by junhyeop         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,8 @@ void	free_token(t_token *token)
 	{
 		rmv = token;
 		token = token->next;
-		free(rmv->cmd);
+		if (rmv->cmd)
+			free(rmv->cmd);
 		free(rmv);
 	}
 }
@@ -30,7 +31,6 @@ void	free_list(t_head *head, char *str)
 	t_list	*tmp;
 	t_list	*rmv;
 
-	free(str);
 	tmp = head->top;
 	while (tmp)
 	{
@@ -44,4 +44,6 @@ void	free_list(t_head *head, char *str)
 	head->size = 0;
 	head->top = NULL;
 	head->get_error = 0;
+	head->processes = NULL;
+	free(str);
 }
